Expanded range for cached and stored-in-db guild member exp (devotion) from uint32 to uint64. (bugreport:3983) Capped value when sending to client to INT_MAX to prevent confusion.
git-svn-id: https://svn.code.sf.net/p/rathena/svn/trunk@14242 54d463be-8e91-2dee-dedb-b68131a5f0ec

+++ sql-files/main.sql
@@ -182,7 +182,7 @@ CREATE TABLE IF NOT EXISTS `guild` (
   `connect_member` tinyint(6) unsigned NOT NULL default '0',
   `max_member` tinyint(6) unsigned NOT NULL default '0',
   `average_lv` smallint(6) unsigned NOT NULL default '1',
-  `exp` int(11) unsigned NOT NULL default '0',
+  `exp` bigint(20) unsigned NOT NULL default '0',
   `next_exp` int(11) unsigned NOT NULL default '0',
   `skill_point` tinyint(11) unsigned NOT NULL default '0',
   `mes1` varchar(60) NOT NULL default '',
